+++ src/relax/vm/naive_allocator.h
@@ -18,44 +18,43 @@
  */
 
 /*!
- * \file tvm/relax/vm/pooled_allocator.h
+ * \file tvm/relax/vm/naive_allocator.h
  */
-#ifndef TVM_RELAX_VM_POOLED_ALLOCATOR_H_
-#define TVM_RELAX_VM_POOLED_ALLOCATOR_H_
+#ifndef TVM_RELAX_VM_NAIVE_ALLOCATOR_H_
+#define TVM_RELAX_VM_NAIVE_ALLOCATOR_H_
 
 #include <tvm/relax/vm/memory_manager.h>
 #include <tvm/runtime/device_api.h>
 
 #include <atomic>
-#include <mutex>
-#include <unordered_map>
-#include <vector>
 
 namespace tvm {
 namespace relax {
 namespace vm {
 
-class PooledAllocator final : public Allocator {
+class NaiveAllocator final : public Allocator {
  public:
-  static constexpr size_t kDefaultPageSize = 4096;
+  explicit NaiveAllocator(Device dev) : Allocator(kNaive), used_memory_(0), device_(dev) {}
 
-  explicit PooledAllocator(Device dev, size_t page_size = kDefaultPageSize)
-      : Allocator(kPooled), page_size_(page_size), used_memory_(0), device_(dev) {}
+  Buffer Alloc(size_t nbytes, size_t alignment, DLDataType type_hint) override {
+    Buffer buf;
+    buf.device = device_;
+    buf.size = nbytes;
+    buf.data =
+        runtime::DeviceAPI::Get(device_)->AllocDataSpace(device_, nbytes, alignment, type_hint);
+    used_memory_.fetch_add(nbytes, std::memory_order_relaxed);
+    DLOG(INFO) << "allocate " << nbytes << " B, used memory " << used_memory_ << " B";
+    return buf;
+  }
 
-  ~PooledAllocator() { ReleaseAll(); }
-
-  Buffer Alloc(size_t nbytes, size_t alignment, DLDataType type_hint) override;
-
-  void Free(const Buffer& buffer) override;
+  void Free(const Buffer& buffer) override {
+    runtime::DeviceAPI::Get(device_)->FreeDataSpace(buffer.device, buffer.data);
+    used_memory_.fetch_sub(buffer.size, std::memory_order_relaxed);
+    DLOG(INFO) << "free " << buffer.size << " B, used memory " << used_memory_ << " B";
+  }
 
  private:
-  void ReleaseAll();
-
- private:
-  size_t page_size_;
   std::atomic<size_t> used_memory_;
-  std::unordered_map<size_t, std::vector<Buffer> > memory_pool_;
-  std::recursive_mutex mu_;
   Device device_;
 };
 
@@ -63,4 +62,4 @@ class PooledAllocator final : public Allocator {
 }  // namespace relax
 }  // namespace tvm
 
-#endif  // TVM_RELAX_VM_POOLED_ALLOCATOR_H_
+#endif  // TVM_RELAX_VM_NAIVE_ALLOCATOR_H_
